Move deploy bot response builder and event handlers into a separate package
