+++ bot/response_builder.go
@@ -1,4 +1,4 @@
-package server
+package bot
 
 import (
 	"fmt"
