+++ bot/slack_topic_manager.go
@@ -1,4 +1,4 @@
-package server
+package bot
 
 import (
 	"log"
